refactor(core): split pipeline entry, facet place, scene, boxplot batches
Extract public API re-exports and run-id allocation; facet grid input and
map/place helpers; scene/finalize input types; boxplot rects vs segments;
candidate locate types and outlier/annotation/ordinal-rank modules; setup
normalize/edition/theme; prepare-panels bin ranges and empty-layer warns;
smooth frame packing; single-panel placement packing; panel layout result
pack. Facades keep import paths stable. Add characterization tests for
run ids and outlier context.
