+++ packages/core/src/pipeline/assemble-scene-build.ts
@@ -1,31 +1,15 @@
 /**
  * Scene assembly from panel placements, axes, and legends.
  */
-import type { GeometryBatch, Scene, SceneLegend } from "../scene.js";
-import type { ThemeTokens } from "../theme.js";
-import type { PositionScale } from "../scales/train.js";
+import type { Scene } from "../scene.js";
 
+import type { AssembleSceneInput } from "./assemble-scene-input.js";
 import { placeSceneLegends } from "./assemble-scene-legends.js";
 import { assembleScenePanels } from "./assemble-scene-panels.js";
-import type { FacetPanelDef } from "./facets.js";
-import type { PanelPlacement } from "./panel-layout.js";
 
-export function assembleScene(input: {
-  width: number;
-  height: number;
-  placements: readonly PanelPlacement[];
-  facetPanels: readonly FacetPanelDef[];
-  displayScales: (p: number) => { h: PositionScale; v: PositionScale };
-  hTitle: string;
-  vTitle: string;
-  batches: GeometryBatch[];
-  legendBlock: { legends: SceneLegend[]; width: number };
-  topBand: number;
-  theme: ThemeTokens;
-  title: string;
-  subtitle: string;
-  caption: string;
-}): Scene {
+export type { AssembleSceneInput } from "./assemble-scene-input.js";
+
+export function assembleScene(input: AssembleSceneInput): Scene {
   const {
     width,
     height,
