+++ packages/core/src/pipeline/build-candidates-datum-locate.ts
@@ -2,7 +2,6 @@
  * Locate frame/batch context for one identity-indexed candidate fact.
  */
 import type { CandidateBuildFacts } from "../candidate-store.js";
-import type { CellValue } from "../table.js";
 
 import type { IdentityCandidateResolveContext } from "./build-candidates-datum-ctx.js";
 import { resolveOutlierContext } from "./build-candidates-datum-context.js";
@@ -10,23 +9,10 @@ import {
   makeSourceValueLookup,
   resolveCandidateFieldChannels,
 } from "./build-candidates-datum-fields.js";
+import type { LocatedIdentityCandidate } from "./build-candidates-datum-locate-types.js";
 import { resolveCandidateFrameRow } from "./build-candidates-frame-row.js";
-import type { LayerFrame } from "./types.js";
 
-export interface LocatedIdentityCandidate {
-  sourceRow: number | null;
-  frame: LayerFrame | undefined;
-  outlierSourceRow: number | null;
-  frameRow: number;
-  derivedGroup: number;
-  sourceValue: (field: string | undefined) => CellValue;
-  xField: string | undefined;
-  yField: string | undefined;
-  colorField: string | undefined;
-  fillField: string | undefined;
-  seriesByRow: Map<string, number>;
-  sourceRowsByGroup: Map<string, number[]>;
-}
+export type { LocatedIdentityCandidate } from "./build-candidates-datum-locate-types.js";
 
 export function locateIdentityCandidate(
   ctx: IdentityCandidateResolveContext,
