+++ packages/core/src/pipeline/finalize-model.ts
@@ -1,30 +1,15 @@
 /**
  * Finalize phase: layer contracts, domains, candidates, and RenderModel.
  */
-import type { PortableSpec } from "@ggsvelte/spec";
-
-import type { Scene } from "../scene.js";
-
 import { assembleFinalizeRenderModel } from "./finalize-model-assemble.js";
 import { buildFinalizeCandidates } from "./finalize-model-candidates.js";
 import { resolveFinalizeContracts } from "./finalize-model-contracts.js";
-import type { PanelLayoutResult } from "./panel-layout.js";
-import type { PreparedPanels } from "./prepare-panels.js";
-import type { TrainedPipelineScales } from "./train-pipeline-scales.js";
-import type { Advisory, PipelineWarning, RenderModel, RunOptions } from "./types.js";
+import type { FinalizeRenderModelInput } from "./finalize-model-input.js";
+import type { RenderModel } from "./types.js";
 
-export function finalizeRenderModel(input: {
-  runId: number;
-  normalized: PortableSpec;
-  options: RunOptions;
-  flip: boolean;
-  prepared: PreparedPanels;
-  trained: TrainedPipelineScales;
-  panelLayout: PanelLayoutResult;
-  scene: Scene;
-  warnings: PipelineWarning[];
-  advisories: Advisory[];
-}): RenderModel {
+export type { FinalizeRenderModelInput } from "./finalize-model-input.js";
+
+export function finalizeRenderModel(input: FinalizeRenderModelInput): RenderModel {
   const {
     runId,
     normalized,
