+++ packages/core/src/pipeline/public-api.ts
@@ -1,0 +1,19 @@
+/**
+ * Public pipeline contract re-exports (import path stability for consumers
+ * of @ggsvelte/core via ./pipeline.ts).
+ */
+export type {
+  Advisory,
+  AxisValueFormatter,
+  LayerBackend,
+  MappedField,
+  NamedData,
+  PipelineWarning,
+  RenderModel,
+  ResolvedColorScale,
+  RunOptions,
+  ScaleDomainSnapshot,
+  TrainedScales,
+} from "./types.js";
+export { CANVAS_AUTO_THRESHOLD, PipelineError } from "./types.js";
+export { batchMarkCount } from "./geometry.js";
