+++ packages/core/src/pipeline/run-id.ts
@@ -1,0 +1,8 @@
+/**
+ * Monotonic pipeline run identity (module-global).
+ */
+let nextRunId = 0;
+
+export function allocatePipelineRunId(): number {
+  return ++nextRunId;
+}
